consumer/requirements.txt
aiokafka==0.12.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic==2.12.5
